update
- add discount
- add discount style
- change column money type (tbStock, tbOrderDT)

+++ server/models/tbStock.js
@@ -5,11 +5,11 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(11, 2),
       allowNull: false,
     },
     discount: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(11, 2),
       allowNull: false,
     },
     discountType: {
